pyproject.toml
[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    #PEP8-Naming
    "N"
]

[tool.poetry]
name = "openeu-backend"
version = "0.0.1"
description = ""
authors = ["CSEE"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.13"
pre-commit = "3.7.1"
python-dotenv = "1.0.1"
fastapi = "0.115.12"
poetry-core = "2.1.3"
supabase = "2.15.1"
mypy = "1.15.0"
requests = "2.32.3"
types-requests = "2.32.0.20250328"
bs4 = "0.0.2"
scrapy = "2.13.0"
parsel = "1.10.0"
deepl = "^1.22.0"

[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"
